requirements.txt
pandas
numpy
matplotlib
mplfinance
TA-Lib
requests
python-telegram-bot==20.4
pytz